+++ bugs/core/world/entities/ant/base/genetic/genes/adjusting_development_appetite_gene.py
@@ -18,7 +18,7 @@ class AdjustingDevelopmentAppetiteGene(BaseGene):
         super().__init__(GenesTypes.ADJUSTING_DEVELOPMENT_APPETITE, ChromosomeTypes.ADJUSTING, domination_code)
 
     def affect(self, phenotype: Phenotype):
-        phenotype.required_food = 2 * (phenotype.strength + phenotype.defense + phenotype.speed/2)
+        phenotype.required_food = int(2 * (phenotype.strength + phenotype.defense + phenotype.speed/2))
 
     def merge(self, another_gene: 'AdjustingDevelopmentAppetiteGene') -> BaseGene:
         dominating_gene = super().merge(another_gene)
